优化MVVM的代码结构

+++ DemoMVVM/DemoMVVM/ModelUserInfo.h
@@ -11,7 +11,7 @@
 @interface ModelUserInfo : NSObject
 
 @property (nonatomic, copy)     NSString    *name;
-@property (nonatomic, assign)   NSInteger   age;
+@property (nonatomic, assign)   int         age;
 @property (nonatomic, copy)     NSString    *city;
 
 @end
